Change gpio name to pinedio. update udev rules.
Changed the gpio name that the device presents.  It was CH341, these can come in
a variety of configurations. Ours is quite specific and deserves a proper name.
This should make the life of application developers much easier.

This also let me clean up the udev rules and not mess up /dev with an
unnecessary symlink.