Imporved udev rules.
updated the udev ruled to bind spidev0.0 when the pinedio is pugged in. the
advantage is that anyone in the "gpio" group can use the device without any
extra privileges. this patch also updates the Makefile to install and activate
the udev rules and remove them and reload the rules during an uninstall.

Signed-off-by: Winford <dwinford@pm.me>